cli: Allow provider names to be passed without "provider" key
This makes for an easier experience when adding a registered context
provider on the command line.

Also refactors some utilities away from the implementation files.
